show details pipeline

+++ src/main/java/com/backend/devsecopsplatform_backend/controller/environment/DeployResponse.java
@@ -15,7 +15,7 @@ public class DeployResponse {
 
     private UUID environmentId;
     private String environmentName;
-    private Integer gitlabPipelineId;
+    private Long gitlabPipelineId;
     private String pipelineStatus;
     private String pipelineWebUrl;
     private String message;
